voice/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.10, <3.15"

dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "firebase-admin>=7.2.0",
    "groq>=1.1.1",
    "httpx>=0.28.1",
    "livekit-agents[silero,turn-detector]~=1.4",
    "livekit-api>=1.1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.4.5",
    "numpy>=2.2.6",
    "python-dotenv",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
